add admin and roles

+++ controller/user.controller.js
@@ -32,7 +32,7 @@ const signup = async (req, res) => {
       return res.status(400).json({ message: error.errors[0].message });
     }
 
-    const { name, email, password, username } = req.body;
+    const { name, email, password, username, role } = req.body;
 
     //Check if user already exists
     const existingUser = await User.findOne({ email });
@@ -46,6 +46,7 @@ const signup = async (req, res) => {
       email,
       password,
       username,
+      role: role || "user",
     });
 
     //Generate JWT token
@@ -53,7 +54,13 @@ const signup = async (req, res) => {
 
     return res.status(201).json({
       message: "User created succesfully",
-      user: { id: user._id, name: user.name },
+      user: {
+        id: user._id,
+        name: user.name,
+        username: user.username,
+        email: user.email,
+        role: user.role,
+      },
       token,
     });
   } catch (error) {
@@ -92,7 +99,13 @@ const login = async (req, res) => {
 
     return res.status(200).json({
       message: "Login successful",
-      user: { id: user._id, name: user.name },
+      user: {
+        id: user._id,
+        name: user.name,
+        username: user.username,
+        email: user.email,
+        role: user.role,
+      },
       token,
     });
   } catch (error) {
@@ -183,9 +196,23 @@ const updateUserById = async (req, res) => {
 const deleteUserById = async (req, res) => {
   try {
     const { userId } = req.params;
+    const currentUserId = req.user._id;
+    const currentUserRole = req.user.role;
 
     if (!mongoose.Types.ObjectId.isValid(userId)) {
       return res.status(404).json({ message: "Invalid User ID" });
+    }
+
+    const user = await User.findById(userId);
+
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    if (currentUserId !== userId && currentUserRole !== "admin") {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to delete the user" });
     }
 
     await User.findByIdAndDelete(userId);
